resolved all routing and refactored email controllers to appropriate controller
