+++ server/routes/analytics-routes.js
@@ -4,12 +4,10 @@ const adminAuth = require('../middleware/adminAuth');
 
 const router = express.Router();
 
-router.use(adminAuth); // Protect all analytics routes
+router.use(adminAuth);
 
 router.get('/course-enrollment-revenue', analyticsControllers.getCourseEnrollmentAndRevenueData);
-
 router.get('/monthly-revenue', analyticsControllers.getMonthlyRevenueData);
-
 router.get('/course-analytics', analyticsControllers.getCourseAnalytics);
 
 module.exports = router;
