+++ server/routes/course-routes.js
@@ -6,9 +6,7 @@ const courseController = require('../controllers/course-controllers');
 const router = express.Router();
 
 router.get('/', courseController.getCourses);
-
 router.get('/:cid', courseController.getCourseById);
-
 router.get('/:cid/modules', courseController.getCourseModules);
 
 router.post(
@@ -16,8 +14,8 @@ router.post(
   [check('title').not().isEmpty(), check('description').not().isEmpty()],
   courseController.createCourse
 );
-router.delete('/:cid', courseController.deleteCourse);
 
+router.delete('/:cid', courseController.deleteCourse);
 router.put('/:cid', courseController.updateCourse);
 
 module.exports = router;
