+++ server/routes/email-routes.js
@@ -4,10 +4,13 @@ const adminAuth = require('../middleware/adminAuth');
 
 const router = express.Router();
 
+// Public routes
+router.post('/forgot-login-code', emailController.forgotLoginCode);
+
+// Protected routes
 router.use(adminAuth);
 
-router.post('/send-student-login-code', emailController.sendStudentLoginCode);
-
-router.post('/resend-certificate', emailController.resendCertificate);
+router.post('/send-login-code', emailController.sendLoginCode);
+router.post('/resend-certificate', emailController.resendCertificate); 
 
 module.exports = router;
